Moving everything into small app dir
