--- app.yaml
@@ -1,21 +1,0 @@
-application: performanceboard-public
-version: 1
-runtime: go
-api_version: go1
-threadsafe: true
-
-handlers:
-- url: /static
-  static_dir: client
-- url: /favicon.ico
-  static_files: client/img/favicon.ico
-  upload: nil
-- url: /.*
-  script: _go_app
-
-skip_files:
-- ^(.*/)?#.*#$
-- ^(.*/)?.*~$
-- ^(.*/)?.*\.py[co]$
-- ^(.*/)?.*/RCS/.*$
-- ^(.*/)?\..*$
