--- client/index.js
@@ -1,5 +1,0 @@
-var angular = require('angular');
-
-angular.element(document).ready(function() {
-    angular.bootstrap(document, [require('./src/app')]);
-});
