+++ client/src/main.js
@@ -1,0 +1,5 @@
+var angular = require('angular');
+
+angular.element(document).ready(function() {
+    angular.bootstrap(document, [require('./app')]);
+});
