--- index.yaml
@@ -1,33 +1,0 @@
-indexes:
-
-# AUTOGENERATED
-
-# This index.yaml is automatically updated whenever the dev_appserver
-# detects that a new type of query is run.  If you want to manage the
-# index.yaml file manually, remove the above marker line (the line
-# saying "# AUTOGENERATED").  If you want to manage some indexes
-# manually, move them above the marker line.  The index.yaml file is
-# automatically uploaded to the admin console when you next deploy
-# your application using appcfg.py.
-
-- kind: AggregateMetric
-  ancestor: yes
-  properties:
-  - name: BinType
-  - name: Namespace
-  - name: StartTime
-    direction: desc
-
-- kind: Metric
-  ancestor: yes
-  properties:
-  - name: Namespace
-  - name: Start
-    direction: desc
-
-- kind: Taxonomy
-  properties:
-  - name: BoardKey
-  - name: Namespace
-  - name: Start
-    direction: desc
