--- server/templates/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>PerformanceBoard</title>
-    <meta charset="UTF8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content="">
-    <link rel="stylesheet" href="/static/build.css">
-    <script src="/static/build.js"></script>
-  </head>
-  <body>
-    <div id="body">
-      <ng-view></ng-view>
-    </div>
-  </body>
-</html>
